back/todoapp/todoapp_api/views.py
from django.shortcuts import render
from rest_framework import viewsets
from .models import User,Task
from .serializers import UserSerializer,TaskSerializer

class UserViewSet(viewsets.ModelViewSet):
    queryset = User.objects.all()
    serializer_class = UserSerializer

class TaskViewSet(viewsets.ModelViewSet):
    queryset = Task.objects.all()
    serializer_class = TaskSerializer